improvements due to patched pwm-ir-tx driver
With the hrtimer-based pwm-ir-tx driver (patched and deployed a custom 6.6.y kernel),
test only transmitting config once, and skip the coarse categorization of pulses and spaces.
With the more precise driver, it should not be necessary to send more than once, and
received messages should also be more precise (when we are detecting our own messages).