Schedule timer jobs to update the Power setting
Instead of trying to second-guess if Power should be on or off at the
current time based on being before or after the timer on and off times,
schedule internal timer jobs that update the power setting in the db at the
same time that the inverter turns on or off. This should avoid wrong
assumptions that occur when power has been manually turned on or off while
timers are enabled.

Note that if the timer settings are changed with paninv_rc, then
paninv_controller will be unaware of the changes and will not reschedule
the timer jobs, until some update is done by paninv_controller via the web
interface or a cron job.